backend/pyproject.toml
[project]
name = "nutrilens-backend"
version = "2.0.0"
description = "AI-powered dual-camera nutrition tracking API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "groq>=0.4.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "scipy>=1.11.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ── UV ────────────────────────────────────────────────────────────────────────
[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations (type hints)
    "S",   # flake8-bandit (security)
    "RUF", # ruff-specific
]
ignore = [
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "ANN401", # dynamically typed Any
    "S101",   # use of assert (fine in tests)
    "B008",   # do not perform function calls in default arguments (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
